Rename items to Postings and and redux-thunk
