+++ client-side/src/components/PostingCard.js
