+++ client-side/src/pages/Favorites.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import ItemCard from '../components/ItemCard';
+import PostingCard from '../components/PostingCard';
 
 function Favorites({ items }) {
   return (
     <div>
       <h1>Favorites</h1>
       {items.map((item) => (
-        <ItemCard key={item.id} item={item} />
+        <PostingCard key={item.id} item={item} />
       ))}
     </div>
   );
