+++ client-side/src/reducers/postings.js
